@@ -1,112 +1,92 @@
 {
-  "name": "pvCostMonitor",
-  "scripts": {
-    "build": "npx projen build",
-    "clobber": "npx projen clobber",
-    "compile": "npx projen compile",
-    "default": "npx projen default",
-    "dev": "npx projen dev",
-    "eject": "npx projen eject",
-    "eslint": "npx projen eslint",
-    "package": "npx projen package",
-    "post-compile": "npx projen post-compile",
-    "post-upgrade": "npx projen post-upgrade",
-    "pre-compile": "npx projen pre-compile",
-    "test": "npx projen test",
-    "test:watch": "npx projen test:watch",
-    "upgrade": "npx projen upgrade",
-    "watch": "npx projen watch",
-    "projen": "npx projen"
-  },
-  "devDependencies": {
+  "name": "pv-manager",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "*",
-    "@types/jest": "^29.4.0",
-    "@types/node": "^14",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.12",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
-    "@typescript-eslint/eslint-plugin": "^5",
-    "@typescript-eslint/parser": "^5",
-    "eslint": "^8",
-    "eslint-import-resolver-node": "^0.3.7",
-    "eslint-import-resolver-typescript": "^3.5.3",
-    "eslint-plugin-import": "^2.27.5",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.4.2",
-    "jest-junit": "^13",
-    "jsdom": "21",
-    "npm-check-updates": "^16",
-    "projen": "^0.67.36",
-    "ts-jest": "^29.0.5",
-    "typescript": "^4.0.3"
-  },
-  "dependencies": {
     "@types/react-datepicker": "^4",
     "axios": "^1",
     "bootstrap": "^5",
     "bootswatch": "^5",
     "chart.js": "^4",
+    "lint-staged": "^13.1.1",
     "moment": "^2",
-    "next": "^13.1.6",
+    "prettier": "^2.8.4",
     "react": "^18.2.0",
     "react-chartjs-2": "^5",
     "react-datepicker": "^4",
     "react-dom": "^18.2.0",
-    "react-scripts": "^5",
-    "web-vitals": "*"
+    "husky": "^8.0.3",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4"
   },
-  "license": "Apache-2.0",
-  "version": "0.0.0",
+  "devDependencies": {
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.3",
+    "prettier": "^2.8.3",
+    "ts-node": "^10.9.1"
+  },
+  "lint-staged": {
+    "{src,tools,acceptance,public}/**/*.{t,j}s{,x}": [
+      "prettier --write",
+      "eslint --fix"
+    ],
+    "src/**/*.scss": [
+      "stylelint --fix"
+    ]
+  },
   "jest": {
+    "resetMocks": false,
+    "collectCoverageFrom": [
+      "src/**/*.{j,t}s{,x}",
+      "!src/index.ts"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 100,
+        "functions": 100,
+        "lines": 100,
+        "statements": 100
+      }
+    },
     "transformIgnorePatterns": [
       "node_modules/(?!axios)"
     ],
     "moduleNameMapper": {
       "\\.(css|sass)$": "identity-obj-proxy"
-    },
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.ts?(x)",
-      "<rootDir>/(test|src)/**/*(*.)@(spec|test).ts?(x)"
-    ],
-    "clearMocks": true,
-    "collectCoverage": true,
-    "coverageReporters": [
-      "json",
-      "lcov",
-      "clover",
-      "cobertura",
-      "text"
-    ],
-    "coverageDirectory": "coverage",
-    "coveragePathIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "watchPathIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "reporters": [
-      "default",
-      [
-        "jest-junit",
-        {
-          "outputDirectory": "test-reports"
-        }
-      ]
-    ],
-    "preset": "ts-jest"
+    }
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --coverage",
+    "test:watch": "npm run test -- --watchAll",
+    "lint": "eslint '{src,tools,acceptance,public}/**/*.{t,j}s{,x}' && stylelint 'src/**/*.{,s}css'",
+    "lint:fix": "eslint --fix '{src,tools,acceptance,public}/**/*.{t,j}s{,x}' && stylelint --fix 'src/**/*.{,s}css'",
+    "pre-commit": "lint-staged",
+    "prepare": "husky install"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
       "react-app/jest"
     ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
   },
   "browserslist": {
     "production": [
